+++ Alien-Invasion/Units/AlienMonster.h
@@ -3,7 +3,7 @@
 class AlienMonster : public Unit
 {
 public:
-	AlienMonster(int H, int P, int AC, int tj, Game* Gameptr);
+	AlienMonster(double H, double P, int AC, int tj, Game* Gameptr);
 	virtual void Attack();
 	~AlienMonster();
 };
