+++ Alien-Invasion/Units/EarthGunnery.cpp
@@ -1,7 +1,7 @@
 #include "EarthGunnery.h"
 #include"../Game class/Game.h"
 
-EarthGunnery::EarthGunnery(int H, int P, int AC, int tj, Game* Gameptr) :Unit(H, P, AC, tj, Gameptr)
+EarthGunnery::EarthGunnery(double H, double P, int AC, int tj, Game* Gameptr) :Unit(H, P, AC, tj, Gameptr)
 {
 	Type = earthGunnery;
 }
@@ -45,7 +45,7 @@ void EarthGunnery::Attack()
 			unit2->setTa(pGame->GetCurrentTime()); //Set Ta (first attacked time)
 
 
-			int Damage = (this->getHealth() * this->getPower() / 100) /
+			double Damage = (this->getHealth() * this->getPower() / 100) /
 				sqrt(unit2->getHealth());	//Damage Formula
