+++ Alien-Invasion/Units/EarthTank.cpp
@@ -2,7 +2,7 @@
 #include <cmath>
 #include"../Game class/Game.h"
 
-EarthTank::EarthTank(int H, int P, int AC, int tj, Game* Gameptr) :Unit(H, P, AC, tj, Gameptr)
+EarthTank::EarthTank(double H, double P, int AC, int tj, Game* Gameptr) :Unit(H, P, AC, tj, Gameptr)
 {
 	Type = earthTank;
 	AttackBoth = false;
@@ -53,7 +53,7 @@ void EarthTank::Attack()
 			unit2->setTa(pGame->GetCurrentTime()); //Set Ta (first attacked time)
 
 
-			int Damage = (this->getHealth() * this->getPower() / 100) /
+			double Damage = (this->getHealth() * this->getPower() / 100) /
 				sqrt(unit2->getHealth());	//Damage Formula
