+++ Alien-Invasion/Units/HealUnit.h
@@ -3,7 +3,7 @@
 class HealUnit :  public Unit
 {
 public:
-	HealUnit(int H, int P, int AC, int tj, Game* Gameptr);
+	HealUnit(double H, double P, int AC, int tj, Game* Gameptr);
 	virtual void Attack();
 	~HealUnit();
 };
